+++ python/GafferImageTest/ImageTestCase.py
@@ -279,7 +279,7 @@ parent["color"] = imath.Color4f( 0.5, 0.6, 0.7, 0.8 ) if context.get( "collect:l
 
 		deep = GafferImage.Empty()
 		node["in"].setInput( deep["out"] )
-		self.assertRaisesRegex( RuntimeError, 'Deep data not supported in input "in*', GafferImage.ImageAlgo.image, node["out"] )
+		self.assertRaisesRegex( RuntimeError, 'Deep data not supported in input "in*', GafferImage.ImageAlgo.tiles, node["out"] )
 
 	@staticmethod
 	def imagesPath() :