+++ python/GafferUI/Image.py
@@ -35,10 +35,8 @@
 ##########################################################################
 
 import os
-import imath
 
 import IECore
-import IECoreImage
 
 import GafferUI
 
@@ -50,7 +48,7 @@ from Qt import QtWidgets
 # as either a filename or an IECore.ImagePrimitive.
 class Image( GafferUI.Widget ) :
 
-	def __init__( self, imagePrimitiveOrFileName, **kw ) :
+	def __init__( self, fileName, **kw ) :
 
 		GafferUI.Widget.__init__( self, QtWidgets.QLabel(), **kw )
 
@@ -58,19 +56,12 @@ class Image( GafferUI.Widget ) :
 		# the same size.
 		self._qtWidget().setSizePolicy( QtWidgets.QSizePolicy( QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed ) )
 
-		self.updateImage( imagePrimitiveOrFileName )
+		self.updateImage( fileName )
 
-	def updateImage( self, imagePrimitiveOrFileName ) :
+	def updateImage( self, fileName ) :
 
-		if isinstance( imagePrimitiveOrFileName, str ) :
-			pixmap = self._qtPixmapFromFile( str( imagePrimitiveOrFileName ) )
-		elif isinstance( imagePrimitiveOrFileName, IECoreImage.ImagePrimitive ) :
-			pixmap = self._qtPixmapFromImagePrimitive( imagePrimitiveOrFileName )
-		else :
-			pixmap = None
-
-		if pixmap is not None :
-			self._qtWidget().setPixmap( pixmap )
+		if fileName is not None :
+			self._qtWidget().setPixmap( self._qtPixmapFromFile( str( fileName ) ) )
 
 		self.__pixmapHighlighted = None
 		self.__pixmapDisabled = None
@@ -222,54 +213,6 @@ class Image( GafferUI.Widget ) :
 		icon = QtGui.QIcon( self._qtPixmapHighlighted() if highlighted else self._qtPixmap() )
 		icon.addPixmap( self._qtPixmapDisabled(), QtGui.QIcon.Disabled )
 		return icon
-
-	## \todo Deprecate and remove - we want to phase out ImagePrimitive.
-	# Although we don't use this function or the `Image( ImagePrimitive )`
-	# constructor in Gaffer itself, we can't do this immediately because some
-	# external code currently depends on it.
-	@staticmethod
-	def _qtPixmapFromImagePrimitive( image ) :
-
-		image = image.copy()
-		IECoreImage.ColorAlgo.transformImage( image, "linear", "sRGB" )
-
-		y = image["Y"] if "Y" in image else None
-		r = image["R"] if "R" in image else None
-		g = image["G"] if "G" in image else None
-		b = image["B"] if "B" in image else None
-
-		if r and g and b :
-			channels = [ r, g, b ]
-		elif y :
-			channels = [ y, y, y ]
-		else :
-			raise ValueError( "Expected RGB or Y channels in image" )
-
-		if "A" in image :
-			channels.reverse()
-			channels.append( image["A"] )
-			format = QtGui.QImage.Format_ARGB32_Premultiplied
-		else :
-			format = QtGui.QImage.Format_RGB888
-
-		interleaved = IECore.DataInterleaveOp()(
-			data = IECore.ObjectVector( channels ),
-			targetType = IECore.UCharVectorData.staticTypeId(),
-		)
-
-		imageSize = image.dataWindow.size() + imath.V2i( 1 )
-
-		s = interleaved.toString()
-		image = QtGui.QImage( s, imageSize.x, imageSize.y, format )
-		# The image is referencing the data we provided directly (s), and
-		# expects us to keep it alive as long as the image/pixmap is
-		# used. Since we don't want to do that, we make a deep copy of
-		# the image so we can dispose of s.
-		image = image.copy()
-
-		pixmap = QtGui.QPixmap( image )
-
-		return pixmap
 
 	__pixmapCache = None
 	@classmethod