+++ Changes.md
@@ -15,6 +15,8 @@ Breaking Changes
   - Removed deprecated support for passing `script` and `context` arguments to `frameRange()` method.
     Use `with script.context()` instead.
 - TractorDispatcher : Removed deprecated support for `preSpoolSignal` slots without `taskData` arguments.
+- ImageAlgo : Removed `image()` method.
+- ImageGadget, Image : Removed constructors taking an `IECoreImage::ImagePrimitive`.
 
 1.7.x.x (relative to 1.7.1.0)
 =======